pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-cloud-storage>=3.1.0",
    "gunicorn>=23.0.0",
    "openai>=1.75.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "pytesseract>=0.3.13",
    "sift-stack-py>=0.5.1",
    "werkzeug>=3.1.3",
    "flask-wtf>=1.2.2",
    "flask-migrate>=4.1.0",
    "requests>=2.32.3",
    "numpy>=2.2.5",
    "sqlalchemy>=2.0.40",
    "python-dotenv>=1.1.0",
    "twilio>=9.5.2",
]
